[flask] 1차 수정
