+++ backend/application/sql/schema.sql
@@ -1,9 +1,9 @@
 CREATE TABLE user_info(
     USER_EMAIL VARCHAR(500) NOT NULL,
     USER_PW VARCHAR(500) NOT NULL,
-    USER_NICKNAME VARCHAR(200) NOT NULL,
-    USER_AGE INT NOT NULL,
-    USER_SEX INT NOT NULL,
-    USER_BIRTH INT NOT NULL,
+    NICKNAME VARCHAR(200) NOT NULL,
+    BIRTH DATE NOT NULL,
+    USER_PROFILE LONGBLOB,
     PRIMARY KEY(USER_EMAIL)
+
 );
